requirements.txt
Django==1.10.6
django-bootstrap3==8.2.1
djangorestframework==3.6.2
wheel==0.24.0
